Migrated Check if Linked List is Palindrome
